+++ utils/gear-hooks/src/hooks/index.ts
@@ -11,6 +11,7 @@ import {
   useIsVoucherExists,
   useVoucherBalance,
   useVoucher,
+  useBalance,
   useBalanceFormat,
   SendMessageOptions,
   UseSendMessageOptions,
@@ -34,6 +35,7 @@ export {
   useIsVoucherExists,
   useVoucherBalance,
   useVoucher,
+  useBalance,
   useBalanceFormat,
   useAccount,
   useAlert,
