+++ utils/gear-hooks/src/index.ts
@@ -11,6 +11,7 @@ import {
   useIsVoucherExists,
   useVoucherBalance,
   useVoucher,
+  useBalance,
   useBalanceFormat,
   useAccount,
   useAlert,
@@ -61,6 +62,7 @@ export {
   useIsVoucherExists,
   useVoucherBalance,
   useVoucher,
+  useBalance,
   useBalanceFormat,
   useAccount,
   useAlert,
